tASK 12jabo

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 int x;
-long long int y, z, sum;
+unsigned long int y, z, sum;
 
 y = 2;
 z = 1;
